+++ packages/blocks-helpers/src/default-attributes/index.ts
@@ -3,6 +3,7 @@ import { AlignAttribute } from "./align";
 import { ClassNameAttribute } from "./className";
 import { LayoutAttribute } from "./layout";
 import { LockAttribute } from "./lock";
+import { MetaDataAttribute } from "./metadata";
 import { StyleAttribute } from "./style";
 
 export type DefaultAttributes<Supports extends BlockSupports> =
@@ -10,4 +11,5 @@ export type DefaultAttributes<Supports extends BlockSupports> =
 		ClassNameAttribute<Supports> &
 		LayoutAttribute<Supports> &
 		LockAttribute<Supports> &
+		MetaDataAttribute<Supports> &
 		StyleAttribute<Supports>;